requirements.txt
Django==5.2.8
djangorestframework==3.16.1
drf-spectacular==0.29.0
djangorestframework-api-key==3.1.0
gunicorn==23.0.0
python-dotenv==1.2.1
pymysql==1.1.1
